deploy-requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==1.0.0; python_version >= '3.6'
aiohttp==3.8.1; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
antlr4-python3-runtime==4.8
anyio==3.5.0; python_full_version >= '3.6.2'
asgiref==3.5.0; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
asynctest==0.13.0; python_version < '3.8'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
click==8.0.4; python_version >= '3.6'
colorlog==6.6.0; python_version >= '3.6'
datasets==1.18.4
dill==0.3.4; python_version >= '2.7' and python_version != '3.0'
fastapi==0.75.0
filelock==3.6.0; python_version >= '3.7'
flatbuffers==2.0
frozenlist==1.3.0; python_version >= '3.7'
fsspec[http]==2022.2.0; python_version >= '3.7'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-auth==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
grpcio==1.44.0; python_version >= '3.6'
h11==0.13.0; python_version >= '3.6'
huggingface-hub==0.4.0; python_version >= '3.6'
hydra-colorlog==1.1.0
hydra-core==1.1.1
idna==3.3; python_version >= '3.5'
importlib-metadata==4.11.2; python_version < '3.8'
importlib-resources==5.4.0; python_version < '3.9'
joblib==1.1.0; python_version >= '3.6'
markdown==3.3.6; python_version >= '3.6'
multidict==6.0.2; python_version >= '3.7'
multiprocess==0.70.12.2
numpy==1.21.5; python_version < '3.11' and python_version >= '3.7'
oauthlib==3.2.0; python_version >= '3.6'
omegaconf==2.1.1
onnxruntime==1.10.0
packaging==21.3; python_version >= '3.6'
pandas==1.3.5; python_full_version >= '3.7.1'
protobuf==3.19.4; python_version >= '3.5'
pyarrow==7.0.0; python_version >= '3.7'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.9.0; python_full_version >= '3.6.1'
pydeprecate==0.3.1; python_version >= '3.6'
pyparsing==3.0.7; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytorch-lightning==1.5.10
pytz==2021.3
pyyaml==6.0; python_version >= '3.6'
regex==2022.3.2; python_version >= '3.6'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
responses==0.18.0; python_version >= '3.7'
rsa==4.8; python_version >= '3.6'
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3; python_version < '3.11' and python_version >= '3.7'
setuptools==59.5.0; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.17.1; python_version >= '3.6'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0; python_version >= '3.6'
threadpoolctl==3.1.0; python_version >= '3.6'
tokenizers==0.11.6
torch==1.11.0; python_version >= '3.7'
torchmetrics==0.7.2; python_version >= '3.6'
tqdm==4.63.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==4.17.0
typing-extensions==4.1.1; python_version < '3.8'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.17.5
werkzeug==2.0.3; python_version >= '3.6'
wheel==0.37.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
xxhash==3.0.0; python_version >= '3.6'
yarl==1.7.2; python_version >= '3.6'
zipp==3.7.0; python_version >= '3.7'
